--- Common/Common.TasksLibrary/Task2/Base/IBuildCalculator.cs
@@ -1,6 +1,0 @@
-﻿namespace Common.TasksLibrary.Task2.Base;
-
-public interface IBuildCalculator
-{
-    public Calculator Build();
-}
