+++ Common/Common.TasksLibrary/Task2/Base/ICalculatorCommand.cs
@@ -1,4 +1,6 @@
-﻿namespace Common.TasksLibrary.Task2.Base;
+﻿using Common.TasksLibrary.Task2.Handlers;
+
+namespace Common.TasksLibrary.Task2.Base;
 
 public interface ICalculatorCommand
 {
