--- Common/Common.TasksLibrary/Task2/Base/ISpecifyLogger.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.Extensions.Logging;
-
-namespace Common.TasksLibrary.Task2.Base;
-
-public interface ISpecifyLogger
-{
-    public ISpecifyOutput LogBy(ILogger logger);
-}
