--- Common/Common.TasksLibrary/Task2/Base/ISpecifyOutput.cs
@@ -1,8 +1,0 @@
-﻿using Common.TasksLibrary.Task2.Output;
-
-namespace Common.TasksLibrary.Task2.Base;
-
-public interface ISpecifyOutput
-{
-    public ISpecifyStorage OutBy(CalculatorOutputOptions options);
-}
