--- Common/Common.TasksLibrary/Task2/Base/ISpecifyStorage.cs
@@ -1,6 +1,0 @@
-﻿namespace Common.TasksLibrary.Task2.Base;
-
-public interface ISpecifyStorage
-{
-    public IBuildCalculator StoreBy(BaseContainer container);
-}
