+++ Common/Common.TasksLibrary/Task2/Calculator.cs
@@ -1,4 +1,6 @@
 ﻿using Common.TasksLibrary.Task2.Exceptions;
+using Common.TasksLibrary.Task2.Factories;
+using Common.TasksLibrary.Task2.Handlers;
 using Microsoft.Extensions.Logging;
 
 namespace Common.TasksLibrary.Task2;
