+++ Common/Common.TasksLibrary/Task2/Commands/DefineCommand.cs
@@ -1,6 +1,7 @@
 ﻿using Common.TasksLibrary.Constants;
 using Common.TasksLibrary.Task2.Base;
 using Common.TasksLibrary.Task2.Exceptions;
+using Common.TasksLibrary.Task2.Handlers;
 
 namespace Common.TasksLibrary.Task2.Commands;
