+++ Common/Common.TasksLibrary/Task2/Commands/MultiplyCommand.cs
@@ -1,5 +1,6 @@
 ﻿using Common.TasksLibrary.Task2.Base;
 using Common.TasksLibrary.Task2.Exceptions;
+using Common.TasksLibrary.Task2.Handlers;
 
 namespace Common.TasksLibrary.Task2.Commands;
