+++ Common/Common.TasksLibrary/Task2/Providers/CalculatorOutputOptions.cs
