+++ Common/Common.TasksLibrary/Task2/Providers/ConsoleOutput.cs
