+++ Tasks/Tasks.Task2/Program.cs
@@ -2,6 +2,8 @@
 using Common.TasksLibrary;
 using Common.TasksLibrary.Extensions;
 using Common.TasksLibrary.Task2;
+using Common.TasksLibrary.Task2.Factories;
+using Common.TasksLibrary.Task2.Handlers;
 using Common.TasksLibrary.Task2.Output;
 using Microsoft.Extensions.Logging;
