+++ Tests/PracticeTests/Task2Tests/CalculatorExecutionFromFileTest.cs
@@ -2,6 +2,8 @@
 using System.Text;
 using Common.TasksLibrary.Constants;
 using Common.TasksLibrary.Task2;
+using Common.TasksLibrary.Task2.Factories;
+using Common.TasksLibrary.Task2.Handlers;
 using Common.TasksLibrary.Task2.Output;
 using Microsoft.Extensions.Logging;
 using PracticeTests.Properties;
