+++ Tests/PracticeTests/Task2Tests/CommandsExecutionTests.cs
@@ -1,6 +1,7 @@
 ﻿using Common.TasksLibrary.Constants;
 using Common.TasksLibrary.Task2;
 using Common.TasksLibrary.Task2.Commands;
+using Common.TasksLibrary.Task2.Handlers;
 using Common.TasksLibrary.Task2.Output;
 
 namespace PracticeTests.Task2Tests;
